# docs/swagger.yaml
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /transfer/{bucketName}/{objectPath}:
    get:
      description: Download a Storj object by its bucketName & objectPath
      parameters:
      - description: Bucket Name
        in: path
        name: bucketName
        required: true
        type: string
      - description: Object Path
        in: path
        name: objectPath
        required: true
        type: string
      produces:
      - application/json
      - text/plain
      - application/octet-stream
      - video/mp4
      - image/png
      - image/jpeg
      - image/gif
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Download a Storj object
      tags:
      - Object Operations
  /transfer/{bucketName}/{objectPathPrefix}:
    post:
      consumes:
      - multipart/form-data
      description: Upload an object at the specified bucketName & objectPathPrefix on Storj
      parameters:
      - description: Bucket Name
        in: path
        name: bucketName
        required: true
        type: string
      - default: /
        description: Object Path Prefix
        in: path
        name: objectPathPrefix
        required: true
        type: string
      - description: File Name
        in: formData
        name: file
        required: true
        type: file
      - default: anonymous
        description: Your name
        in: formData
        name: name
        type: string
      - default: anonymous
        description: Your email
        in: formData
        name: email
        type: string
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Upload an object to Storj
      tags:
      - Object Operations
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
x-extension-openapi:
  example: value on a json format